requirements.txt
alembic==1.0.7
backcall==0.1.0
Click==7.0
decorator==4.3.2
Django==2.1.7
django-autofixture==0.12.1
Faker==0.9.1
Flask==1.0.2
Flask-Admin==1.5.3
flask-marshmallow==0.8.0
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
infinity==1.4
intervals==0.8.1
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.13.6
marshmallow-sqlalchemy==0.13.2
mimesis==3.1.0
mixer==6.1.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.18
SQLAlchemy-Utils==0.33.11
text-unidecode==1.2
traitlets==4.3.2
validators==0.12.4
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
WTForms-Alchemy==0.16.8
WTForms-Components==0.10.3
